requirements.txt
backtrader
#yfinance[nospam,repair]
tqdm
scikit-learn
tensorflow[and-cuda]

pandas
numpy
python-dateutil
requests
tenacity
aiohttp
aiolimiter
aiofiles
sentencepiece